Horizontal overflow without javascript
It seems that the same layout can be accomplished without javascript by setting the works div to display:flex.
It might be that what Benedikt had in mind was to scroll the window horizontal when the user scrolls vertical, in which case we will probably need to reach for the js again....

ref. #4